requirements.txt
flask
pandas
google-api-python-client
google-auth
google-auth-httplib2
pillow
pytest
ruff
